refactor: 팀 컨벤션에 맞게 커뮤니티 피드, 커뮤니티 상세 보기, 글 작성 화면 라우팅 구조 리팩토링
