+++ web/src/pages/communityDetail/CommunityDetail.tsx
@@ -1,7 +1,3 @@
-import React from "react";
-
-const CommunityDetail = () => {
+export const CommunityDetail = () => {
   return <div>CommunityDetail</div>;
 };
-
-export default CommunityDetail;
