+++ web/src/pages/communityWrite/CommunityWrite.tsx
@@ -1,7 +1,3 @@
-import React from "react";
-
-const CommunityWrite = () => {
+export const CommunityWrite = () => {
   return <div>CommunityWrite</div>;
 };
-
-export default CommunityWrite;
