+++ web/src/pages/index.ts
@@ -1,1 +1,4 @@
 export { Home } from "./home/Home";
+export { CommunityDetail } from "./communityDetail/CommunityDetail";
+export { CommunityFeed } from "./communityFeed/CommunityFeed";
+export { CommunityWrite } from "./communityWrite/CommunityWrite";
